display skeleton while waiting for project stats api response
