+++ frontend/src/components/pages/admin/DashboardProjectStorage.tsx
@@ -1,78 +1,34 @@
 import axios, { AxiosResponse } from 'axios';
-import { useMemo, useState } from 'react';
-import { useLoaderData } from 'react-router-dom';
+import { Suspense } from 'react';
+import { Await, defer, useLoaderData } from 'react-router-dom';
+import { ExclamationTriangleIcon } from '@heroicons/react/24/outline';
 
+import DashboardProjectStorageTable from './DashboardProjectStorageTable';
 import { ProjectStatistics } from './DashboardTypes';
-import Pagination from '../../Pagination';
+import DashboardProjectStorageTableSkeleton from './DashboardProjectStorageTableSkeleton';
 
 export async function loader() {
-  try {
-    const response: AxiosResponse<ProjectStatistics[]> = await axios.get(
-      `${import.meta.env.VITE_API_V1_STR}/admin/project_statistics`
-    );
-    if (response.status === 200) {
-      return response.data;
-    } else {
-      return [];
-    }
-  } catch (_err) {
-    return [];
-  }
+  const response: Promise<AxiosResponse<ProjectStatistics[]>> = axios.get(
+    `${import.meta.env.VITE_API_V1_STR}/admin/project_statistics`
+  );
+  return defer({ response });
+}
+
+function ErrorElement() {
+  return (
+    <div className="h-full w-full flex flex-col items-center justify-center">
+      <ExclamationTriangleIcon className="h-10 w-10 text-red-600" />
+      <span className="text-2xl font-semibold text-primary">
+        Error occurred. Unable to load user project statistics.
+      </span>
+    </div>
+  );
 }
 
 export default function DashboardProjectStorage() {
-  const userProjectStats = useLoaderData() as ProjectStatistics[];
-
-  const [currentPage, setCurrentPage] = useState(0);
-
-  const MAX_ITEMS = 5; // max number of users per page
-  const TOTAL_PAGES = Math.ceil(userProjectStats.length / MAX_ITEMS);
-
-  /**
-   * Updates the current selected pagination page.
-   * @param newPage Index of new page.
-   */
-  function updateCurrentPage(newPage: number): void {
-    const total_pages = Math.ceil(
-      userProjectStats ? userProjectStats.length : 0 / MAX_ITEMS
-    );
-
-    if (newPage + 1 > total_pages) {
-      setCurrentPage(total_pages - 1);
-    } else if (newPage < 0) {
-      setCurrentPage(0);
-    } else {
-      setCurrentPage(newPage);
-    }
-  }
-
-  /**
-   * Filters users by search text and limits to current page.
-   * @param users Users to filter.
-   * @returns
-   */
-  function filterAndSlice(stats: ProjectStatistics[]) {
-    return stats.slice(currentPage * MAX_ITEMS, MAX_ITEMS + currentPage * MAX_ITEMS);
-  }
-
-  /**
-   * Returns available users on page limitations.
-   * @param users Users to filter, limit, and sort.
-   * @returns Array of filtered users.
-   */
-  function getAvailableUsers(stats): ProjectStatistics[] {
-    return filterAndSlice(stats);
-  }
-
-  const userProjectStatsSorted = useMemo(() => {
-    return [...getAvailableUsers(userProjectStats)].sort(
-      (a, b) => b.total_storage - a.total_storage
-    );
-  }, [userProjectStats]);
-
-  if (userProjectStats.length === 0) {
-    return <section className="w-full bg-white">No data</section>;
-  }
+  const projectStatisticsApiResponse = useLoaderData() as {
+    response: Promise<ProjectStatistics[]>;
+  };
 
   return (
     <section className="w-full bg-white">
@@ -81,34 +37,20 @@ export default function DashboardProjectStorage() {
           <h2 className="text-3xl font-bold text-gray-900 sm:text-4xl">
             User Project Storage
           </h2>
-
           <p className="mt-4 text-gray-500 sm:text-xl">
             Summary of total disk usage by user based on project ownership.
           </p>
+          <Suspense fallback={<DashboardProjectStorageTableSkeleton />}>
+            <Await
+              resolve={projectStatisticsApiResponse.response}
+              errorElement={<ErrorElement />}
+            >
+              {(response) => (
+                <DashboardProjectStorageTable userProjectStats={response.data} />
+              )}
+            </Await>
+          </Suspense>
         </div>
-        <table className="relative w-full border-separate border-spacing-y-1 border-spacing-x-1">
-          <thead>
-            <tr className="h-12 sticky top-0 text-slate-700 bg-slate-300">
-              <th className="p-4">Name</th>
-              <th className="p-4">Total projects</th>
-              <th className="p-4">Total storage (GB)</th>
-            </tr>
-          </thead>
-          <tbody className="max-h-96 overflow-y-auto">
-            {userProjectStatsSorted.map((stats) => (
-              <tr key={stats.id} className="text-center">
-                <td className="p-4 bg-slate-100">{stats.user}</td>
-                <td className="p-4 bg-white">{stats.total_projects}</td>
-                <td>{(stats.total_storage / 1024 ** 3).toFixed(3)}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-        <Pagination
-          currentPage={currentPage}
-          updateCurrentPage={updateCurrentPage}
-          totalPages={TOTAL_PAGES}
-        />
       </div>
     </section>
   );
